fix: Make uni-migration CI green by extending log/test renames downstream
Why: PR #1634 broke CI in two ways. (1) The Scala 3.8 bump made companion
class constructors private, so wvlet-lang's TreeNodeCompat fell through
to its newInstance reflection path and threw IllegalAccessException for
133 RoundTrip tests. (2) Every module that consumes wvlet-lang
(wvlet-runner, wvlet-cli, wvlet-server, wvlet-spec, wvlet-ui, ...) still
imported wvlet.log.* while their inherited base classes now mix in
wvlet.uni.log.LogSupport — a LazyLogger mixin clash that blocked the
runner / cli / Scala.js packaging jobs.

Fixes:
- wvlet-lang/.jvm TreeNodeCompat: switch from Reflect.companionOf
  (annotation-required) to Reflect.companionOfUnchecked (JVM-only escape
  hatch in uni 2026.1.4 specifically intended for this case). This is
  what airframe.surface.reflect.ReflectTypeUtil.companionObject did.
- Mechanical wvlet.log.* -> wvlet.uni.log.* rename across wvlet-runner,
  wvlet-cli, wvlet-server, wvlet-spec, wvlet-ui*, wvlet-labs, wvc,
  wvlet-sdk-js, vscode-wvlet (~33 files).
- Mechanical wvlet.airspec.AirSpec -> wvlet.uni.test.UniTest in tests
  except for 7 files that rely on AirSpec-only DI features (initDesign,
  constructor injection, inCI). Those keep AirSpec for now; UniTest does
  not yet expose equivalent DI hooks.
- wvlet-spec/TDTrinoSpec: bridge AirSpec's wvlet.log.LogLevel to
  wvlet.uni.log.LogLevel via name (`uni.log.LogLevel(logger.getLogLevel.name)`).
- wvlet-ui (parent ui module): add explicit uni dep — it doesn't
  transitively pull in uni and now needs LogSupport from there.
- Drop two more dead imports surfaced while migrating:
  wvlet-runner/InMemoryExecutor (wvlet.log.io.IOUtil), and the LogTimestampFormatter
  rename in wvlet-ui-main/ConsoleLogWindow.

Verified locally:
- ./sbt projectJVM/Test/compile: green
- ./sbt projectJS/Test/compile: green
- ./sbt projectNative/Test/compile: green
- ./sbt langJVM/testOnly *RoundTripSpecBasic: 123/123 (was 0/123)

wvlet-labs has a pre-existing bug (CompilerOptions.apply does not have
a `phases` parameter) that surfaced when this branch tried to compile it
locally; labs is not part of CI's project graph so it does not block the
build.
