+++ wvlet-cli/src/main/scala/wvlet/lang/cli/Clipboard.scala
@@ -15,7 +15,7 @@ package wvlet.lang.cli
 
 import wvlet.lang.compiler.OS
 import wvlet.lang.runner.ProcessUtil
-import wvlet.log.LogSupport
+import wvlet.uni.log.LogSupport
 
 import java.io.BufferedWriter
 import java.io.OutputStreamWriter
