+++ wvlet-runner/src/main/scala/wvlet/lang/runner/WvletScriptRunner.scala
@@ -23,9 +23,9 @@ import wvlet.lang.catalog.Profile
 import wvlet.lang.compiler.*
 import wvlet.lang.compiler.query.QueryProgressMonitor
 import wvlet.lang.runner.*
-import wvlet.log.LogRotationHandler
-import wvlet.log.LogSupport
-import wvlet.log.Logger
+import wvlet.uni.log.LogRotationHandler
+import wvlet.uni.log.LogSupport
+import wvlet.uni.log.Logger
 
 import java.io.BufferedWriter
 import java.io.FilterOutputStream
