+++ wvlet-runner/src/main/scala/wvlet/lang/runner/connector/QueryProgressMonitor.scala
@@ -1,7 +1,7 @@
 package wvlet.lang.runner.connector
 
 import io.trino.jdbc.QueryStats
-import wvlet.log.LogSupport
+import wvlet.uni.log.LogSupport
 import wvlet.lang.compiler.query.QueryMetric
 
 case class TrinoQueryMetric(stats: QueryStats) extends QueryMetric
