+++ wvlet-server/src/main/scala/wvlet/lang/server/StaticContentApi.scala
@@ -3,7 +3,7 @@ package wvlet.lang.server
 import wvlet.airframe.http.Endpoint
 import wvlet.airframe.http.HttpMessage
 import wvlet.airframe.http.StaticContent
-import wvlet.log.LogSupport
+import wvlet.uni.log.LogSupport
 
 import java.io.File
