+++ wvlet-ui-playground/src/main/scala/wvlet/lang/ui/playground/PlaygroundUI.scala
@@ -2,7 +2,7 @@ package wvlet.lang.ui.playground
 
 import wvlet.airframe.Design
 import wvlet.airframe.rx.html.RxElement
-import wvlet.log.LogSupport
+import wvlet.uni.log.LogSupport
 import wvlet.airframe.rx.html.all.*
 import wvlet.lang.ui.component.monaco.EditorBase
 import wvlet.lang.ui.component.Icon
